// packages/types/src/interfaces/metadata/types.ts
export interface CustomMetadata15 extends Struct {
  readonly map: BTreeMap<Text, CustomValueMetadata15>;
}

/** @name CustomValueMetadata15 */
export interface CustomValueMetadata15 extends Struct {
  readonly type: SiLookupTypeId;
  readonly value: Bytes;
}

/** @name EnumDeprecationInfoV16 */
export interface EnumDeprecationInfoV16 extends BTreeMap<u8, VariantDeprecationInfoV16> {}
// ...
export interface ExtrinsicMetadataV15 extends Struct {
  readonly version: u8;
  readonly addressType: SiLookupTypeId;
  readonly callType: SiLookupTypeId;
  readonly signatureType: SiLookupTypeId;
  readonly extraType: SiLookupTypeId;
  readonly signedExtensions: Vec<SignedExtensionMetadataV14>;
}

/** @name ExtrinsicMetadataV16 */
export interface ExtrinsicMetadataV16 extends Struct {
  readonly versions: Bytes;
  readonly addressType: SiLookupTypeId;
  readonly callType: SiLookupTypeId;
  readonly signatureType: SiLookupTypeId;
  readonly transactionExtensionsByVersion: BTreeMap<u8, Vec<Compact<u32>>>;
  readonly transactionExtensions: Vec<TransactionExtensionMetadataV16>;
}
// ...
export interface ItemDeprecationInfoV16 extends Enum {
  readonly isNotDeprecated: boolean;
  readonly isDeprecatedWithoutNote: boolean;
  readonly isDeprecated: boolean;
  readonly asDeprecated: {
    readonly note: Text;
    readonly since: Option<Text>;
  } & Struct;
  readonly type: 'NotDeprecated' | 'DeprecatedWithoutNote' | 'Deprecated';
}
// ...
export interface MetadataV15 extends Struct {
  readonly lookup: PortableRegistry;
  readonly pallets: Vec<PezpalletMetadataV15>;
  readonly extrinsic: ExtrinsicMetadataV15;
  readonly type: SiLookupTypeId;
  readonly apis: Vec<RuntimeApiMetadataV15>;
  readonly outerEnums: OuterEnums15;
  readonly custom: CustomMetadata15;
}

/** @name MetadataV16 */
export interface MetadataV16 extends Struct {
  readonly lookup: PortableRegistry;
  readonly pallets: Vec<PezpalletMetadataV16>;
  readonly extrinsic: ExtrinsicMetadataV16;
  readonly apis: Vec<RuntimeApiMetadataV16>;
  readonly outerEnums: OuterEnums15;
  readonly custom: CustomMetadata15;
}
// ...
export interface OuterEnums15 extends Struct {
  readonly callType: SiLookupTypeId;
  readonly eventType: SiLookupTypeId;
  readonly errorType: SiLookupTypeId;
}

/** @name PezpalletAssociatedTypeMetadataV16 */
export interface PezpalletAssociatedTypeMetadataV16 extends Struct {
  readonly name: Text;
  readonly type: SiLookupTypeId;
  readonly docs: Vec<Text>;
}

/** @name PezpalletCallMetadataLatest */
export interface PezpalletCallMetadataLatest extends PezpalletCallMetadataV16 {}

/** @name PezpalletCallMetadataV14 */
export interface PezpalletCallMetadataV14 extends Struct {
  readonly type: SiLookupTypeId;
}

/** @name PezpalletCallMetadataV16 */
export interface PezpalletCallMetadataV16 extends Struct {
  readonly type: SiLookupTypeId;
  readonly deprecationInfo: EnumDeprecationInfoV16;
}

/** @name PezpalletConstantMetadataLatest */
export interface PezpalletConstantMetadataLatest extends PezpalletConstantMetadataV16 {}

/** @name PezpalletConstantMetadataV14 */
export interface PezpalletConstantMetadataV14 extends Struct {
  readonly name: Text;
  readonly type: SiLookupTypeId;
  readonly value: Bytes;
  readonly docs: Vec<Text>;
}

/** @name PezpalletConstantMetadataV16 */
export interface PezpalletConstantMetadataV16 extends Struct {
  readonly name: Text;
  readonly type: SiLookupTypeId;
  readonly value: Bytes;
  readonly docs: Vec<Text>;
  readonly deprecationInfo: ItemDeprecationInfoV16;
}

/** @name PezpalletErrorMetadataLatest */
export interface PezpalletErrorMetadataLatest extends PezpalletErrorMetadataV16 {}

/** @name PezpalletErrorMetadataV14 */
export interface PezpalletErrorMetadataV14 extends Struct {
  readonly type: SiLookupTypeId;
}

/** @name PezpalletErrorMetadataV16 */
export interface PezpalletErrorMetadataV16 extends Struct {
  readonly type: SiLookupTypeId;
  readonly deprecationInfo: EnumDeprecationInfoV16;
}

/** @name PezpalletEventMetadataLatest */
export interface PezpalletEventMetadataLatest extends PezpalletEventMetadataV16 {}

/** @name PezpalletEventMetadataV14 */
export interface PezpalletEventMetadataV14 extends Struct {
  readonly type: SiLookupTypeId;
}

/** @name PezpalletEventMetadataV16 */
export interface PezpalletEventMetadataV16 extends Struct {
  readonly type: SiLookupTypeId;
  readonly deprecationInfo: EnumDeprecationInfoV16;
}
// ...
export interface PezpalletMetadataV15 extends Struct {
  readonly name: Text;
  readonly storage: Option<PezpalletStorageMetadataV14>;
  readonly calls: Option<PezpalletCallMetadataV14>;
  readonly events: Option<PezpalletEventMetadataV14>;
  readonly constants: Vec<PezpalletConstantMetadataV14>;
  readonly errors: Option<PezpalletErrorMetadataV14>;
  readonly index: u8;
  readonly docs: Vec<Text>;
}

/** @name PezpalletMetadataV16 */
export interface PezpalletMetadataV16 extends Struct {
  readonly name: Text;
  readonly storage: Option<PezpalletStorageMetadataV16>;
  readonly calls: Option<PezpalletCallMetadataV16>;
  readonly events: Option<PezpalletEventMetadataV16>;
  readonly constants: Vec<PezpalletConstantMetadataV16>;
  readonly errors: Option<PezpalletErrorMetadataV16>;
  readonly associatedTypes: Vec<PezpalletAssociatedTypeMetadataV16>;
  readonly viewFunctions: Vec<PezpalletViewFunctionMetadataV16>;
  readonly index: u8;
  readonly docs: Vec<Text>;
  readonly deprecationInfo: ItemDeprecationInfoV16;
}

/** @name PezpalletStorageMetadataLatest */
export interface PezpalletStorageMetadataLatest extends PezpalletStorageMetadataV16 {}

/** @name PezpalletStorageMetadataV14 */
export interface PezpalletStorageMetadataV14 extends Struct {
  readonly prefix: Text;
  readonly items: Vec<StorageEntryMetadataV14>;
}

/** @name PezpalletStorageMetadataV16 */
export interface PezpalletStorageMetadataV16 extends Struct {
  readonly prefix: Text;
  readonly items: Vec<StorageEntryMetadataV16>;
}

/** @name PezpalletViewFunctionMetadataV16 */
export interface PezpalletViewFunctionMetadataV16 extends Struct {
  readonly id: U8aFixed;
  readonly name: Text;
  readonly inputs: Vec<RuntimeApiMethodParamMetadataV15>;
  readonly output: SiLookupTypeId;
  readonly docs: Vec<Text>;
  readonly deprecationInfo: ItemDeprecationInfoV16;
}

/** @name PortableType */
export interface PortableType extends PortableTypeV14 {}

/** @name PortableTypeV14 */
export interface PortableTypeV14 extends Struct {
  readonly id: Si1LookupTypeId;
  readonly type: Si1Type;
}

/** @name RuntimeApiMetadataLatest */
export interface RuntimeApiMetadataLatest extends RuntimeApiMetadataV16 {}

/** @name RuntimeApiMetadataV15 */
export interface RuntimeApiMetadataV15 extends Struct {
  readonly name: Text;
  readonly methods: Vec<RuntimeApiMethodMetadataV15>;
  readonly docs: Vec<Text>;
}

/** @name RuntimeApiMetadataV16 */
export interface RuntimeApiMetadataV16 extends Struct {
  readonly name: Text;
  readonly methods: Vec<RuntimeApiMethodMetadataV16>;
  readonly docs: Vec<Text>;
  readonly version: Compact<u32>;
  readonly deprecationInfo: ItemDeprecationInfoV16;
}

/** @name RuntimeApiMethodMetadataV15 */
export interface RuntimeApiMethodMetadataV15 extends Struct {
  readonly name: Text;
  readonly inputs: Vec<RuntimeApiMethodParamMetadataV15>;
  readonly output: SiLookupTypeId;
  readonly docs: Vec<Text>;
}

/** @name RuntimeApiMethodMetadataV16 */
export interface RuntimeApiMethodMetadataV16 extends Struct {
  readonly name: Text;
  readonly inputs: Vec<RuntimeApiMethodParamMetadataV15>;
  readonly output: SiLookupTypeId;
  readonly docs: Vec<Text>;
  readonly deprecationInfo: ItemDeprecationInfoV16;
}

/** @name RuntimeApiMethodParamMetadataV15 */
export interface RuntimeApiMethodParamMetadataV15 extends Struct {
  readonly name: Text;
  readonly type: SiLookupTypeId;
}

/** @name SignedExtensionMetadataLatest */
export interface SignedExtensionMetadataLatest extends SignedExtensionMetadataV14 {}

/** @name SignedExtensionMetadataV14 */
export interface SignedExtensionMetadataV14 extends Struct {
  readonly identifier: Text;
  readonly type: SiLookupTypeId;
  readonly additionalSigned: SiLookupTypeId;
}
// ...
export interface StorageEntryMetadataV14 extends Struct {
  readonly name: Text;
  readonly modifier: StorageEntryModifierV14;
  readonly type: StorageEntryTypeV14;
  readonly fallback: Bytes;
  readonly docs: Vec<Text>;
}

/** @name StorageEntryMetadataV16 */
export interface StorageEntryMetadataV16 extends Struct {
  readonly name: Text;
  readonly modifier: StorageEntryModifierV14;
  readonly type: StorageEntryTypeV14;
  readonly fallback: Bytes;
  readonly docs: Vec<Text>;
  readonly deprecationInfo: ItemDeprecationInfoV16;
}
// ...
export interface StorageEntryModifierV13 extends StorageEntryModifierV12 {}

/** @name StorageEntryModifierV14 */
export interface StorageEntryModifierV14 extends StorageEntryModifierV13 {}
// ...
export interface StorageEntryTypeV14 extends Enum {
  readonly isPlain: boolean;
  readonly asPlain: SiLookupTypeId;
  readonly isMap: boolean;
  readonly asMap: {
    readonly hashers: Vec<StorageHasherV14>;
    readonly key: SiLookupTypeId;
    readonly value: SiLookupTypeId;
  } & Struct;
  readonly type: 'Plain' | 'Map';
}
// ...
export interface StorageHasherV14 extends StorageHasherV13 {}
// ...
export interface TransactionExtensionMetadataV16 extends Struct {
  readonly identifier: Text;
  readonly type: SiLookupTypeId;
  readonly implicit: SiLookupTypeId;
}

/** @name VariantDeprecationInfoV16 */
export interface VariantDeprecationInfoV16 extends Enum {
  readonly isDummyVariant: boolean;
  readonly isDeprecatedWithoutNote: boolean;
  readonly isDeprecated: boolean;
  readonly asDeprecated: {
    readonly note: Text;
    readonly since: Option<Text>;
  } & Struct;
  readonly type: 'DummyVariant' | 'DeprecatedWithoutNote' | 'Deprecated';
}